refactor: move Error type out from nested class
